changed initial values
